% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{EDA}
\alias{EDA}
\title{Exploratory Data Analysis Function}
\usage{
EDA(data, select_cols, ...)
}
\arguments{
\item{data}{The dataframe which stores the dataset from where the required columns are selected and filtered for EDA}

\item{select_cols}{Enter the columns that need to be selected from the data eg:- c(c("col1", "col2"))}

\item{...}{It takes in all the columns that need to be filtered with its filtering technique as seperate arguments eg:- col1 > 1, col2 > 5}
}
\value{
Returns the dataframe after selecting the required columns and filtering them based on the criteria given
}
\description{
This function takes in a dataframe and only selects the columns passed in as selects and filter the columns entered to filtered.
}
