% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset.R
\name{read_dataset}
\alias{read_dataset}
\title{Read Dataset Function}
\usage{
read_dataset(path, out_dir)
}
\arguments{
\item{path}{The path where the dataset is which needs to be stored in the given path}

\item{out_dir}{The destination where the entered path for the dataset needs to be saved}
}
\value{
Returns the dataset as a datframe
}
\description{
This function reads the dataset when a path is entered for the dataset and saves it at the destination entered
}
\examples{
path <- "https://github.com/rehan13/ownprac/blob/main/testsheet.csv"
out_dir <- tempfile()
dataset <- read_dataset(path, out_dir)
}
